\name{new_polyglot_label}
\alias{new_polyglot_label}
\alias{.Last.lib}
\title{Create a Polyglot Label}
\usage{
new_polyglot_label(value)

.Last.lib(libpath)
}
\arguments{
  \item{value}{A value coercible to character.}
  \item{libpath}{The library tree from which the package is detached.}
}
\value{An S3 object of class \code{polyglot_label}.}
\description{Creates the package's minimal S3 value.}
